requirements.txt
numpy
pandas
streamlit
xgboost
scikit-learn
imblearn
pillow
